requirements.txt
pandas==2.1.0
numpy==1.26.0
scikit-learn==1.3.0
matplotlib==3.8.0
seaborn==0.13.0
mlxtend==0.23.0
spotipy==2.23.0
sqlalchemy==2.0.0
pyyaml==6.0
jupyter==1.0.0
plotly==5.17.0
networkx==3.2
statsmodels==0.14.0
tqdm==4.66.0
